📖 [RB-102] Add plugins to the Mongoose
Write plugins for soft deleting data from the db and removing
versionKey from response.
