+++ src/server/api/v1/about/model/index.js
@@ -11,10 +11,6 @@ const AboutSchema = new Schema(
       type: String,
       default: '',
     },
-    deleted_at: {
-      type: Date,
-      default: null,
-    },
   },
   {
     timestamps: {
@@ -23,17 +19,5 @@ const AboutSchema = new Schema(
     },
   },
 );
-
-AboutSchema.set('toJSON', {
-  virtuals: true,
-  versionKey: false,
-  transform: (_, obj) => {
-    const updatedObj = { ...obj };
-
-    delete updatedObj._id;
-
-    return updatedObj;
-  },
-});
 
 export default model('About', AboutSchema);
