+++ src/server/api/v1/faqs/model/index.js
@@ -22,10 +22,6 @@ const FAQSchema = new Schema(
       type: Boolean,
       required: true,
     },
-    deleted_at: {
-      type: Date,
-      default: null,
-    },
   },
   {
     timestamps: {
@@ -34,17 +30,5 @@ const FAQSchema = new Schema(
     },
   },
 );
-
-FAQSchema.set('toJSON', {
-  virtuals: true,
-  versionKey: false,
-  transform: (_, obj) => {
-    const updatedObj = { ...obj };
-
-    delete updatedObj._id;
-
-    return updatedObj;
-  },
-});
 
 export default model('FAQ', FAQSchema);
