+++ src/server/api/v1/roles/controller/index.js
@@ -161,17 +161,16 @@ export const deleteRoles = async (req, res, next) => {
       return responseWithError(404, 'Nie znaleziono żadnej roli użytkownika.');
     }
 
-    const updatedRoles = await Role.updateMany(
-      { deleted_at: null, can_be_modified: true, type: 'SUPERUSER' },
-      { deleted_at: Date.now() },
+    const ids = roles.map(({ _id }) => _id.toString());
+
+    const updatedRoles = await Role.softDeleteMany(
+      { _id: { $in: ids }, deleted_at: null },
       { new: true },
     );
 
     if (!updatedRoles) {
       return responseWithError(409, 'Nie udało się usunąć roli użytkownika.');
     }
-
-    const ids = roles.map(({ _id }) => _id.toString());
 
     const deletedRoles = await Role.find({
       _id: { $in: ids },
@@ -212,11 +211,10 @@ export const deleteRole = async (req, res, next) => {
       return responseWithError(409, 'Nie można usunąć tej roli użytkownika.');
     }
 
-    const deletedRole = await Role.findOneAndUpdate(
+    const deletedRole = await Role.findOneAndSoftDelete(
       {
         _id: id, deleted_at: null, can_be_modified: true, type: 'SUPERUSER',
       },
-      { deleted_at: Date.now() },
       { new: true },
     );
