+++ src/server/api/v1/termsOfUse/model/index.js
@@ -10,10 +10,6 @@ const TermsOfUseSchema = new Schema(
       type: String,
       required: true,
     },
-    deleted_at: {
-      type: Date,
-      default: null,
-    },
   },
   {
     timestamps: {
@@ -22,17 +18,5 @@ const TermsOfUseSchema = new Schema(
     },
   },
 );
-
-TermsOfUseSchema.set('toJSON', {
-  virtuals: true,
-  versionKey: false,
-  transform: (_, obj) => {
-    const updatedObj = { ...obj };
-
-    delete updatedObj._id;
-
-    return updatedObj;
-  },
-});
 
 export default model('TermsOfUse', TermsOfUseSchema);
